+++ __main__.py
@@ -1,0 +1,10 @@
+import tkinter as tk
+import os as os
+
+from Keystone.Windows.BindFileEditorWindow import BindFileEditorWindow
+
+
+if (__name__ == "__main__"):
+    win = BindFileEditorWindow()
+
+    tk.mainloop()
